fix: Player FSM 버그 수정
- CurrentState 변수 수정을 안해서 생긴 버그

+++ Assets/Script/Skill/Data/ActiveSkillDataBase.cs
@@ -22,7 +22,7 @@ public class ActiveSkillDataBase : ScriptableObject
             var values = data["skill_value"].ToString().Split(',');
             foreach (var value in values)
             {
-                if(int.TryParse(value, out var result))
+                if(float.TryParse(value, out var result))
                 {
                     activeSkillData.AddValue(result);
                 }
